requirements.txt
fastapi==0.105.0
uvicorn==0.25.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
boto3==1.34.7
black==23.12.1
flake8==6.1.0
pytest==7.4.3
httpx==0.26.0
python-dotenv
PyJWT=2.8
werkzeug=3.0.1